requirements.txt
blinker==1.9.0
certifi==2024.12.14
click==8.1.7
cloudinary==1.41.0
colorama==0.4.6
dnspython==2.7.0
Flask==3.1.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
pymongo==4.10.1
python-dotenv==1.0.1
six==1.17.0
urllib3==2.3.0
Werkzeug==3.1.3
